+++ examples/experiments/tesselating_boxes.py
@@ -6,7 +6,7 @@ import tracemalloc
 import trimesh
 from OCC.Core.TopoDS import TopoDS_Shape
 
-from ada.core.utils import create_guid
+from ada.core.guid import create_guid
 from ada.geom import Geometry
 from ada.geom.solids import Box
 from ada.occ.geom import geom_to_occ_geom
