+++ src/ada/api/presentation_layers.py
@@ -5,7 +5,7 @@ from typing import TYPE_CHECKING
 
 from ada.base.changes import ChangeAction
 from ada.config import logger
-from ada.core.utils import create_guid
+from ada.core.guid import create_guid
 
 if TYPE_CHECKING:
     from ada.base.physical_objects import BackendGeom
