+++ src/ada/cadit/ifc/write/write_stru_components.py
@@ -1,4 +1,4 @@
-from ada.core.utils import create_guid
+from ada.core.guid import create_guid
 
 
 def write_door(f, owner_history, insert_placement, insert_shape):
