+++ src/ada/cadit/ifc/write/write_wall.py
@@ -8,7 +8,7 @@ from ada.cadit.ifc.utils import (
     tesselate_shape,
 )
 from ada.core.constants import O, X, Z
-from ada.core.utils import create_guid
+from ada.core.guid import create_guid
 
 from .write_stru_components import write_door, write_window
