+++ src/ada/cadit/step/stp_to_ifc.py
@@ -11,7 +11,7 @@ from ada.cadit.ifc.utils import add_colour, create_local_placement
 from ada.cadit.ifc.write.write_ifc import IfcWriter
 from ada.cadit.step.store import StepStore
 from ada.config import logger
-from ada.core.utils import create_guid
+from ada.core.guid import create_guid
 from ada.occ.serializers import serialize_shape
