+++ src/ada/fem/formats/code_aster/write/write_nodes.py
@@ -1,58 +1,6 @@
 from typing import TYPE_CHECKING
 
-import numpy as np
-
-from ada.config import Settings
-
-from .write_sets import _add_node_sets
-
 if TYPE_CHECKING:
-    from ada import Part
+    pass
 
 
-def _write_nodes(part: "Part", time_step, profile, families):
-    """
-
-    TODO: Go through each data group and set in HDF5 file and make sure that it writes what was read 1:1.
-        Use cylinder.med as a benchmark.
-
-    Add the following datasets ['COO', 'FAM', 'NUM'] to the 'NOE' group
-
-    :param part:
-    :param time_step:
-    :param profile:
-    :return:
-    """
-    points = np.zeros((int(part.fem.nodes.max_nid), 3))
-
-    def pmap(n):
-        points[int(n.id - 1)] = n.p
-
-    list(map(pmap, part.fem.nodes))
-
-    # Try this
-    if Settings.ca_experimental_id_numbering is True:
-        points = np.array([n.p for n in part.fem.nodes])
-
-    nodes_group = time_step.create_group("NOE")
-    nodes_group.attrs.create("CGT", 1)
-    nodes_group.attrs.create("CGS", 1)
-
-    nodes_group.attrs.create("PFL", np.string_(profile))
-    coo = nodes_group.create_dataset("COO", data=points.flatten(order="F"))
-    coo.attrs.create("CGT", 1)
-    coo.attrs.create("NBR", len(points))
-
-    if Settings.ca_experimental_id_numbering is True:
-        node_ids = [n.id for n in part.fem.nodes]
-        num = nodes_group.create_dataset("NUM", data=node_ids)
-        num.attrs.create("CGT", 1)
-        num.attrs.create("NBR", len(points))
-
-    if len(part.fem.nsets.keys()) > 0:
-        _add_node_sets(nodes_group, part, points, families)
-
-    assembly = part.get_assembly()
-
-    if assembly != part and len(assembly.fem.nsets.keys()) > 0:
-        _add_node_sets(nodes_group, assembly, points, families)
