+++ src/ada/fem/formats/code_aster/write/writer.py
@@ -12,9 +12,8 @@ from ada.fem.utils import is_quad8_shell_elem, is_tri6_shell_elem
 from ..compatibility import check_compatibility
 from .templates import el_convert_str, main_comm_str
 from .write_bc import create_bc_str
-from .write_elements import elements_str
 from .write_materials import materials_str
-from .write_nodes import _write_nodes
+from .write_med import med_nodes, med_elements
 from .write_sections import create_sections_str
 from .write_steps import create_step_str
 
@@ -149,10 +148,10 @@ def write_to_med(name, part: Part, analysis_dir):
         part.fem.sets.add_references()
 
         # Nodes and node sets
-        _write_nodes(part, time_step, profile, families)
+        med_nodes(part, time_step, profile, families)
 
         # Elements (mailles in French) and element sets
-        elements_str(part, time_step, profile, families)
+        med_elements(part, time_step, profile, families)
 
 
 def _write_mesh_presets(f, mesh_name):
